requirements.txt
python-telegram-bot==21.9
google-generativeai==0.8.3
aiohttp==3.11.11
aiosqlite==0.20.0
python-dotenv==1.0.1